+++ src/styles/components/_widget.scss
@@ -3,11 +3,19 @@
 }
 
 .c-widget {
-  height: 250px;
   padding: ($space-1 * 4) ($space-1 * 2) 0;
 
-  > div {
-    height: 100%;
+  .widget-header {
+    margin: 0 0 ($space-1 * 2);
+
+    h2 {
+      font-size: $font-size-default;
+      color: $color-tertiary;
+    }
+  }
+
+  .widget-content {
+    height: 250px;
     background: white;
     box-shadow: 0 1px 0 0 #dbdfe5;
   }
